refactor: Move minimum occurrences from retention conditions to retention options
