+++ frontend/src/scenes/insights/EditorFilters/RetentionOptions.tsx
@@ -3,8 +3,16 @@ import { Link } from '@posthog/lemon-ui'
 import { RetentionCumulativeButton } from '../filters/RetentionCumulativeButton'
 import { RetentionMeanDropdown } from '../filters/RetentionMeanDropdown'
 import { RetentionReferencePicker } from '../filters/RetentionReferencePicker'
+import { MinimumOccurrencesInput } from '../filters/MinimumOccurrencesInput'
+import { retentionLogic } from 'scenes/retention/retentionLogic'
+import { useValues } from 'kea'
+import { insightLogic } from 'scenes/insights/insightLogic'
 
 export function RetentionOptions(): JSX.Element {
+    const { insightProps } = useValues(insightLogic)
+    const { retentionFilter } = useValues(retentionLogic(insightProps))
+    const { minimumOccurrences = 1 } = retentionFilter || {}
+
     return (
         <div className="deprecated-space-y-3" data-attr="retention-options">
             <div className="flex items-center gap-2">
@@ -12,7 +20,11 @@ export function RetentionOptions(): JSX.Element {
                 <RetentionReferencePicker />
             </div>
             <div className="flex items-center gap-2">
-                <div>When users return</div>
+                <div>When users return at least</div>
+                <MinimumOccurrencesInput />
+                <div>time{minimumOccurrences === 1 ? '' : 's'}</div>
+            </div>
+            <div className="flex items-center gap-2">
                 <RetentionCumulativeButton />
                 <div>the period</div>
             </div>
